+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Package - Add contributors, keywords, homepage...
 - Package - Remove an unused package (logger).
 - Dependencies - Freeze the dependencies versions to reduce packages versions changes between projects/environments.
+- Configuration - Rename secret values to envSecret and authSecret.
 
 ### Fixed
 - Integrations - Remove some unnecessary routes.